+++ App.tsx
@@ -5,7 +5,7 @@
  * @format
  */
 
-import { Col, Row, Span } from '@design/StyleAsProps';
+import { Col, Row, Span } from '@design/styleAsProps';
 import React from 'react';
 import type { PropsWithChildren } from 'react';
 import { SafeAreaView, ScrollView, StatusBar, StyleSheet, Text, useColorScheme, View } from 'react-native';
