+++ src/Design/StyleAsProps/Col.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Div, DivPropsType } from '../StyleAsProps';
+import { Div, DivPropsType } from './index';
 
 export const Col = (props: DivPropsType) => {
   return <Div {...props} flexColumn />;
