+++ src/Design/StyleAsProps/Row.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Div, DivPropsType } from '../StyleAsProps';
+import { Div, DivPropsType } from './index';
 
 export const Row = (props: DivPropsType) => {
   return <Div {...props} flexRow />;
